@@ -5,7 +5,7 @@ homepage: https://github.com/kmorkos/flutter_photokit
 version: 0.1.1
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -15,7 +15,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  path_provider: 0.4.1
+  path_provider: 2.0.1
 
 flutter:
   plugin:
